finish implementing sprint race logic
